dashboard/pyproject.toml
[project]
name = "whatsapp-dashboard"
version = "1.0.0"
description = "WhatsApp bridge monitoring dashboard"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "aiosqlite>=0.20",
    "httpx>=0.27",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9",
]

[project.scripts]
whatsapp-dashboard = "whatsapp_dashboard.app:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whatsapp_dashboard"]
